Java HashMap: Считаем прогулы Refactoring
