+++ src/main/java/faang/school/godbless/bjs219358/Main.java
@@ -1,4 +1,4 @@
-package faang.school.godbless.bjs218291;
+package faang.school.godbless.bjs219358;
 
 import java.util.ArrayList;
 import java.util.HashMap;
